vcom class voestalpine
A  vcom class Vehicle Communication System C-601005, use it for simulator application with pymodbus or socket to push data inside registers. Perfomance not tested yet  (may be improved). Version 1.0beta

"The communication is bidirectional – from the vehicle to the receiver and vice versa. The activation signal for
the transponder is emitted by the receiver. The data are transmitted via a Loop Transceiver to the
voestalpine VCOM unit or from the unit to the vehicle"